Resolve make error
When installing using `python setup.py develop`, you would get an error saying that `README.rst` cannot be found. So I updated the program to look for `README.md`. This seems like a bit of a weak fix. Maybe change the `read()` call so that if it can't find the file, it doesn't crash?